Remove dynamic configuration of active listen modes.
When removing active listen modes, the original listen
mode configuration is stored. If there is no new LLCP
activation within a configured time interval (currently
5s), the listen mode configuration is restored to the
old value.

The configuration that is restored is by then not
necessarily the correct configuration anymore. In
particular, this may disturb eSE access which disables
discovery completely.

Disable this at least temporarily to verify that this
is causing crashes on Manta.

Bug: 8775164
Change-Id: I926535a73466c857f7a0f2cac912b5ab1f8c38e1