Rename Broadcom HAL.
Currently it's built using $(TARGET_DEVICE),
but the HAL doesn't have any dependencies on the
target device at all - it's the same for any device.

Renamed to nfc_nci.bcm2079x.default

Also, added a new configuration variable to determine
the name of the HAL, and use that for loading the correct one.

Bug: 16542494
Change-Id: I3c148b1c9491de828030b8d2352d7cd6ed294fce